5.16. Функции возвращаются
