+++ index.html
@@ -233,7 +233,7 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-    <script class="js-functions" src="./js/functions.js"></script>
-    <script class="main-js" src="./js/main.js"></script>
+    <script type="module" src="./js/main.js"></script>
+    <script type="text/javascript" src="./js/functions.js"></script>
 </body>
 </html>
